# api/openapi.yaml
# API contract — owned by THIS repo (contract-per-repo). Phase 2a: real domain.
openapi: 3.0.3
info:
  title: "Savings Account — BIAN Semantic API"
  description: >
    BIAN Service Domain 'Savings Account' (Fulfill) on the 'Savings Account
    Facility Fulfillment Arrangement' control record. Deep implementation:
    no overdraft, min-balance floor, monthly withdrawal cap, daily interest
    accrual + capitalization, KYC gating. Money is integer minor units;
    rates are basis points.
  version: 0.2.0
servers:
  - url: /
  - url: /sd-savings-account
    description: through the platform gateway

paths:
  /v1/service-domain:
    get:
      operationId: getServiceDomain
      summary: Service domain metadata
      responses:
        "200": { description: SD identity and BIAN classification }

  /v1/savings-account-facility-fulfillment-arrangement/initiate:
    post:
      operationId: openAccount
      summary: "Initiate — open a savings account (starts PENDING_KYC)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customerReference]
              properties:
                customerReference: { type: string }
                currency: { type: string, pattern: "^[A-Z]{3}$", default: INR }
                interestRateBp: { type: integer, minimum: 0, maximum: 2000, default: 350,
                                  description: "basis points p.a. (350 = 3.50%)" }
                minBalanceMinor: { type: integer, format: int64, minimum: 0, default: 0 }
      responses:
        "201":
          description: Account opened
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SavingsAccount" }
        "400": { $ref: "#/components/responses/Invalid" }

  /v1/savings-account-facility-fulfillment-arrangement:
    get:
      operationId: listAccounts
      summary: Retrieve (collection)
      responses:
        "200":
          description: All accounts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SavingsAccount" }

  /v1/savings-account-facility-fulfillment-arrangement/{accountId}/retrieve:
    get:
      operationId: retrieveAccount
      summary: Retrieve
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      responses:
        "200":
          description: The account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SavingsAccount" }
        "404": { $ref: "#/components/responses/NotFound" }

  /v1/savings-account-facility-fulfillment-arrangement/{accountId}/control:
    put:
      operationId: controlAccount
      summary: "Control — block | unblock | terminate (close needs balance 0 AND accrued 0)"
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action: { type: string, enum: [block, unblock, terminate] }
      responses:
        "200": { description: New account state }
        "400": { $ref: "#/components/responses/Invalid" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/RuleViolation" }

  /v1/savings-account-facility-fulfillment-arrangement/{accountId}/kyc-result:
    put:
      operationId: applyKycResult
      summary: "KYC outcome callback (Know Your Customer SD)"
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [approved]
              properties:
                approved: { type: boolean }
                reason: { type: string }
      responses:
        "200": { description: ACTIVE on approval, REJECTED on rejection }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/RuleViolation" }

  /v1/savings-account-facility-fulfillment-arrangement/{accountId}/payments/deposit:
    post:
      operationId: deposit
      summary: "Payments BQ — credit funds (never capped)"
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      requestBody: { $ref: "#/components/requestBodies/Posting" }
      responses:
        "201": { $ref: "#/components/responses/Posted" }
        "400": { $ref: "#/components/responses/Invalid" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/RuleViolation" }

  /v1/savings-account-facility-fulfillment-arrangement/{accountId}/payments/withdraw:
    post:
      operationId: withdraw
      summary: "Payments BQ — debit funds (min-balance floor + monthly cap enforced)"
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      requestBody: { $ref: "#/components/requestBodies/Posting" }
      responses:
        "201": { $ref: "#/components/responses/Posted" }
        "400": { $ref: "#/components/responses/Invalid" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/RuleViolation" }

  /v1/savings-account-facility-fulfillment-arrangement/{accountId}/payments/cheque-credit:
    post:
      operationId: chequeCredit
      summary: "Inbound credit from Cheque Processing on cheque.cleared"
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      requestBody: { $ref: "#/components/requestBodies/Posting" }
      responses:
        "201": { $ref: "#/components/responses/Posted" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/RuleViolation" }

  /v1/savings-account-facility-fulfillment-arrangement/{accountId}/payments:
    get:
      operationId: listTransactions
      summary: Transaction history
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      responses:
        "200":
          description: Postings, oldest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Transaction" }
        "404": { $ref: "#/components/responses/NotFound" }

  /v1/savings-account-facility-fulfillment-arrangement/{accountId}/interest/accrue:
    post:
      operationId: accrueInterest
      summary: "Interest BQ — accrue one day of simple interest (floor arithmetic)"
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      responses:
        "200": { description: Account with updated accruedInterestMinor }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/RuleViolation" }

  /v1/savings-account-facility-fulfillment-arrangement/{accountId}/interest/capitalize:
    post:
      operationId: capitalizeInterest
      summary: "Interest BQ — move accrued interest into the balance (INTEREST posting)"
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      responses:
        "201": { $ref: "#/components/responses/Posted" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/RuleViolation" }

  /v1/savings-account-facility-fulfillment-arrangement/{accountId}/balance:
    get:
      operationId: getBalance
      summary: Balance view incl. accrued interest + withdrawals used this month
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      responses:
        "200": { description: Balance view }
        "404": { $ref: "#/components/responses/NotFound" }

components:
  parameters:
    AccountId:
      name: accountId
      in: path
      required: true
      schema: { type: string }

  requestBodies:
    Posting:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [amountMinor]
            properties:
              amountMinor: { type: integer, format: int64, minimum: 1 }
              reference: { type: string }

  responses:
    Posted:
      description: Transaction posted
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Transaction" }
    NotFound:
      description: Unknown account
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RuleViolation:
      description: Business rule forbids this (cap reached, min-balance breach, KYC pending…)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Invalid:
      description: Bad input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }

  schemas:
    SavingsAccount:
      type: object
      properties:
        accountId: { type: string, example: "SA-…" }
        customerReference: { type: string }
        currency: { type: string, example: INR }
        balanceMinor: { type: integer, format: int64, minimum: 0 }
        accruedInterestMinor: { type: integer, format: int64, minimum: 0 }
        interestRateBp: { type: integer }
        minBalanceMinor: { type: integer, format: int64 }
        status:
          type: string
          enum: [PENDING_KYC, ACTIVE, BLOCKED, REJECTED, CLOSED]
        openedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    Transaction:
      type: object
      properties:
        transactionId: { type: string }
        accountId: { type: string }
        type: { type: string, enum: [DEPOSIT, WITHDRAWAL, INTEREST, CHEQUE_CREDIT] }
        amountMinor: { type: integer, format: int64 }
        balanceAfterMinor: { type: integer, format: int64 }
        reference: { type: string }
        postedAt: { type: string, format: date-time }

    Error:
      type: object
      properties:
        code: { type: string, example: WITHDRAWAL_CAP_REACHED }
        message: { type: string }